# backend/openapi/traceability-foss-backend.json
{"openapi":"3.0.1","info":{"title":"Trace-FOSS - OpenAPI Documentation","description":"Trace-FOSS is a system for tracking parts along the supply chain. A high level of transparency across the supplier network enables faster intervention based on a recorded event in the supply chain. This saves costs by seamlessly tracking parts and creates trust through clearly defined and secure data access by the companies and persons involved in the process.","license":{"name":"License: Apache 2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:9999/api","description":"Generated server url"}],"paths":{"/investigations":{"post":{"tags":["Investigations"],"summary":"Start investigations by part ids","description":"The endpoint starts investigations based on part ids provided.","operationId":"investigateAssets","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartInvestigationRequest"}}},"required":true},"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartInvestigationResponse"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartInvestigationResponse"}}}},"201":{"description":"Created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartInvestigationResponse"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/{investigationId}/update":{"post":{"tags":["Investigations"],"summary":"Update investigations by id","description":"The endpoint updates investigations by their id.","operationId":"updateInvestigation","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInvestigationRequest"}}},"required":true},"responses":{"401":{"description":"Authorization failed."},"403":{"description":"Forbidden."},"204":{"description":"No content."}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/{investigationId}/close":{"post":{"tags":["Investigations"],"summary":"Close investigations by id","description":"The endpoint closes investigations by their id.","operationId":"closeInvestigation","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloseInvestigationRequest"}}},"required":true},"responses":{"401":{"description":"Authorization failed."},"403":{"description":"Forbidden."},"204":{"description":"No content."}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/{investigationId}/cancel":{"post":{"tags":["Investigations"],"summary":"Cancles investigations by id","description":"The endpoint cancles investigations by their id.","operationId":"cancelInvestigation","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Authorization failed."},"403":{"description":"Forbidden."},"204":{"description":"No content."}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/{investigationId}/approve":{"post":{"tags":["Investigations"],"summary":"Approves investigations by id","description":"The endpoint approves investigations by their id.","operationId":"approveInvestigation","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Authorization failed."},"403":{"description":"Forbidden."},"204":{"description":"No content."}},"security":[{"oAuth2":["profile email"]}]}},"/edc/notification/contract":{"post":{"tags":["Notifications"],"summary":"Triggers EDC notification contract","description":"The endpoint Triggers EDC notification contract based on notification type and method","operationId":"createNotificationContract","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationContractRequest"}}},"required":true},"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationContractResponse"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationContractResponse"}}}},"201":{"description":"Created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationContractResponse"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets/sync":{"post":{"tags":["Assets"],"summary":"Synchronizes assets from IRS","description":"The endpoint synchronizes the assets from irs.","operationId":"sync","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SyncAssets"}}},"required":true},"responses":{"401":{"description":"Authorization failed."},"403":{"description":"Forbidden."},"201":{"description":"Created."}},"security":[{"oAuth2":["profile email"]}]}},"/assets/detail-information":{"post":{"tags":["Assets"],"summary":"Searches for assets by ids.","description":"The endpoint searchs for assets by id and returns a list of them.","operationId":"getDetailInformation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDetailInformationRequest"}}},"required":true},"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Asset"}}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Asset"}}}}},"200":{"description":"Returns list of found assets","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets/{assetId}":{"get":{"tags":["Assets"],"summary":"Get asset by id","description":"The endpoint returns an asset filtered by id .","operationId":"asset","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}},"200":{"description":"Returns the assets found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}}},"security":[{"oAuth2":["profile email"]}]},"patch":{"tags":["Assets"],"summary":"Updates asset","description":"The endpoint updates asset by provided quality type.","operationId":"updateAsset","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAsset"}}},"required":true},"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}},"200":{"description":"Returns the updated asset","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/registry/reload":{"get":{"tags":["Registry"],"summary":"Triggers reload of shell descriptors","description":"The endpoint Triggers reload of shell descriptors.","operationId":"reload","responses":{"401":{"description":"Authorization failed."},"403":{"description":"Forbidden."},"201":{"description":"Created."}},"security":[{"oAuth2":["profile email"]}]}},"/metrics/registry-lookup":{"get":{"tags":["Registry"],"summary":"Gets Metrics","description":"The endpoint gets metrics for database.","operationId":"metrics","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultRegistryLookupMetric"}}}},"200":{"description":"OK.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultRegistryLookupMetric"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultRegistryLookupMetric"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/{investigationId}":{"get":{"tags":["Investigations"],"summary":"Gets investigations by id","description":"The endpoint returns investigations as paged result by their id.","operationId":"getInvestigation","parameters":[{"name":"investigationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestigationData"}}}},"200":{"description":"OK.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestigationData"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvestigationData"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/received":{"get":{"tags":["Investigations"],"summary":"Gets received investigations","description":"The endpoint returns received investigations as paged result.","operationId":"getReceivedInvestigations","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultInvestigationData"}}}},"200":{"description":"OK.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultInvestigationData"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultInvestigationData"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/investigations/created":{"get":{"tags":["Investigations"],"summary":"Gets created investigations","description":"The endpoint returns created investigations as paged result.","operationId":"getCreatedInvestigations","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultInvestigationData"}}}},"200":{"description":"OK.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultInvestigationData"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultInvestigationData"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/dashboard":{"get":{"tags":["Dashboard"],"summary":"Returns dashboard related data","description":"The endpoint can return limited data based on the user role","operationId":"dashboard","parameters":[{"name":"jwtAuthentication","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtAuthentication"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}},"200":{"description":"Returns dashboard data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets":{"get":{"tags":["Assets"],"summary":"Get assets by pagination","description":"The endpoint returns a paged result of assets.","operationId":"assets","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}},"200":{"description":"Returns the paged result found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets/{assetId}/children/{childId}":{"get":{"tags":["Assets"],"summary":"Get asset by child id","description":"The endpoint returns an asset filtered by child id.","operationId":"asset_1","parameters":[{"name":"assetId","in":"path","required":true,"schema":{"type":"string"}},{"name":"childId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}},"200":{"description":"Returns the asset by childId","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets/supplier":{"get":{"tags":["Assets"],"summary":"Get supplier assets by pagination","description":"The endpoint returns a paged result of supplier assets.","operationId":"supplierAssets","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}},"200":{"description":"Returns the paged result found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets/my":{"get":{"tags":["Assets"],"summary":"Get own assets by pagination","description":"The endpoint returns a paged result of own assets.","operationId":"ownAssets","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}},"200":{"description":"Returns the paged result found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResultAsset"}}}}},"security":[{"oAuth2":["profile email"]}]}},"/assets/countries":{"get":{"tags":["Assets"],"summary":"Get map of assets","description":"The endpoint returns a map for assets consumed by the map.","operationId":"assetsCountryMap","responses":{"401":{"description":"Authorization failed.","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}},"200":{"description":"Returns the assets found","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}},"403":{"description":"Forbidden.","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}}}},"security":[{"oAuth2":["profile email"]}]}}},"components":{"schemas":{"StartInvestigationRequest":{"required":["severity"],"type":"object","properties":{"partIds":{"maxItems":100,"minItems":1,"type":"array","items":{"type":"string"}},"description":{"maxLength":1000,"minLength":15,"type":"string"},"targetDate":{"type":"string","format":"date-time"},"severity":{"type":"string"}}},"StartInvestigationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"UpdateInvestigationRequest":{"required":["status"],"type":"object","properties":{"status":{"type":"string","enum":["CREATED","SENT","RECEIVED","ACKNOWLEDGED","ACCEPTED","DECLINED","CANCELED","CLOSED"]},"reason":{"type":"string"}}},"CloseInvestigationRequest":{"type":"object","properties":{"reason":{"maxLength":1000,"minLength":15,"type":"string"}}},"CreateNotificationContractRequest":{"required":["notificationMethod","notificationType"],"type":"object","properties":{"notificationType":{"type":"string","enum":["QUALITY_INVESTIGATION","QUALITY_ALERT"]},"notificationMethod":{"type":"string","enum":["RECEIVE","UPDATE","RESOLVE"]}}},"CreateNotificationContractResponse":{"type":"object","properties":{"notificationAssetId":{"type":"string"},"accessPolicyId":{"type":"string"},"contractDefinitionId":{"type":"string"}}},"SyncAssets":{"type":"object","properties":{"globalAssetIds":{"maxItems":100,"minItems":1,"type":"array","items":{"type":"string"}}}},"GetDetailInformationRequest":{"type":"object","properties":{"assetIds":{"maxItems":50,"minItems":1,"type":"array","items":{"type":"string"}}}},"Asset":{"type":"object","properties":{"id":{"type":"string"},"idShort":{"type":"string"},"nameAtManufacturer":{"type":"string"},"manufacturerPartId":{"type":"string"},"partInstanceId":{"type":"string"},"manufacturerId":{"type":"string"},"batchId":{"type":"string"},"manufacturerName":{"type":"string"},"nameAtCustomer":{"type":"string"},"customerPartId":{"type":"string"},"manufacturingDate":{"type":"string","format":"date-time"},"manufacturingCountry":{"type":"string"},"supplierPart":{"type":"boolean"},"childDescriptions":{"type":"array","items":{"$ref":"#/components/schemas/ChildDescriptions"}},"underInvestigation":{"type":"boolean"},"qualityType":{"type":"string","enum":["Ok","Minor","Major","Critical","LifeThreatening"]},"van":{"type":"string"}}},"ChildDescriptions":{"type":"object","properties":{"id":{"type":"string"},"idShort":{"type":"string"}}},"UpdateAsset":{"required":["qualityType"],"type":"object","properties":{"qualityType":{"type":"string","enum":["Ok","Minor","Major","Critical","LifeThreatening"]}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageResultRegistryLookupMetric":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/RegistryLookupMetric"}},"page":{"type":"integer","format":"int32"},"pageCount":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalItems":{"type":"integer","format":"int64"}}},"RegistryLookupMetric":{"type":"object","properties":{"startDate":{"type":"string","format":"date-time"},"registryLookupStatus":{"type":"string","enum":["SUCCESSFUL","PARTIALLY_SUCCESS","ERROR"]},"successShellDescriptorsFetchCount":{"type":"integer","format":"int64"},"failedShellDescriptorsFetchCount":{"type":"integer","format":"int64"},"shellDescriptorsFetchDelta":{"type":"integer","format":"int64"},"endDate":{"type":"string","format":"date-time"}}},"InvestigationData":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"description":{"type":"string"},"createdBy":{"type":"string"},"createdDate":{"type":"string"},"assetIds":{"type":"array","items":{"type":"string"}},"channel":{"type":"string","enum":["SENDER","RECEIVER"]},"reason":{"$ref":"#/components/schemas/InvestigationReason"},"sendTo":{"type":"string"},"severity":{"type":"string"}}},"InvestigationReason":{"type":"object","properties":{"close":{"type":"string"},"accept":{"type":"string"},"decline":{"type":"string"}}},"PageResultInvestigationData":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/InvestigationData"}},"page":{"type":"integer","format":"int32"},"pageCount":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalItems":{"type":"integer","format":"int64"}}},"JwtAuthentication":{"type":"object"},"Dashboard":{"type":"object","properties":{"myItems":{"type":"integer","format":"int64"},"otherParts":{"type":"integer","format":"int64"},"investigations":{"type":"integer","format":"int64"}}},"PageResultAsset":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Asset"}},"page":{"type":"integer","format":"int32"},"pageCount":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalItems":{"type":"integer","format":"int64"}}}},"responses":{"badRequestAPI":{"description":"Bad request","content":{"application/json":{"examples":{"default":{"value":"{\"code\" : 400, \"Status\" : \"Bad request\", \"Message\" : \"Bad request\"}"}}}}},"internalServerErrorAPI":{"description":"Internal server error","content":{"application/json":{"examples":{"default":{"value":"{\"code\" : 500, \"Status\" : \"Internal server error\", \"Message\" : \"Internal server error\"}"}}}}}}}}